♻️ (examples): Move instructions to other readmes & fix main readme example
