+++ examples/Readme.md
@@ -1,0 +1,4 @@
+# `gfn` examples
+
+* [scripts/](./scripts/) contains example training scripts for you to play around with different loses and configurations
+* [envs/](./envs/) contains instructions and examples of how to implement your own environments
