+++ src/gfn/utils/__init__.py
@@ -1,3 +1,3 @@
-from .common import add_root_to_path
+from .misc import add_root_to_path
 from .estimators import DiscretePBEstimator, DiscretePFEstimator
 from .modules import NeuralNet, Tabular
