+++ src/gfn/utils/common.py
@@ -1,6 +1,4 @@
-import sys
 from collections import Counter
-from pathlib import Path
 from typing import Dict, Optional
 
 import torch
@@ -104,17 +102,3 @@ def validate(
     if logZ is not None:
         validation_info["logZ_diff"] = abs(logZ - true_logZ)
     return validation_info
-
-
-def get_root() -> Path:
-    """
-    Returns the root directory of the project.
-    """
-    return Path(__file__).resolve().parent.parent.parent.parent
-
-
-def add_root_to_path():
-    """
-    Adds the root directory of the project to the python path.
-    """
-    sys.path.append(str(get_root()))
